:recycle: ReferenceCard 드롭다운 리팩터링
